pyproject.toml
[project]
name = "paralimni"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioredis>=2.0.1",
    "bittensor>=9.3.0",
    "python-decouple>=3.8",
    "uvicorn>=0.34.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
]
